+++ packages/moonstone/CHANGELOG.md
@@ -25,6 +25,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/MoonstoneDecorator` to optimize localized font loading performance
 - `moonstone/VirtualList.VirtualList` and `moonstone/VirtualList.VirtualGridList` to give initial focus
 - `moonstone/GridListImageItem` to show proper `placeholder` and `selectionOverlay`
+- `moonstone/VirtualList.VirtualList` and `moonstone/VirtualList.VirtualGridList` to have the default value for `dataSize`, `pageScroll`, and `spacing` props
 
 ### Changed